packages/backend/requirements.txt
# Core Framework
fastapi==0.109.1
uvicorn[standard]==0.25.0

# Auth & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
cryptography>=42.0.5

# Config & Utilities
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Database & Supabase
supabase==2.4.2
duckdb==0.9.2

# Async & Networking
httpx==0.26.0

# Monitoring & Logging
loguru==0.7.2
